+++ Roslyn~/Coimbra.Roslyn/JetBrainsTypes.cs
@@ -1,0 +1,7 @@
+﻿namespace Coimbra.Roslyn
+{
+    public sealed class JetBrainsTypes
+    {
+        public static readonly TypeString NotNullAttribute = new("NotNull", "JetBrains.Annotations");
+    }
+}