+++ Roslyn~/Coimbra.Roslyn/UnityTypes.cs
@@ -8,6 +8,12 @@
 
         public static readonly TypeString GameObjectClass = new("GameObject", Namespace);
 
+        public static readonly TypeString GuiContentClass = new("GUIContent", Namespace);
+
         public static readonly TypeString ObjectClass = new("Object", Namespace);
+
+        public static readonly TypeString RuntimeInitializeLoadTypeEnum = new("RuntimeInitializeLoadType", Namespace);
+
+        public static readonly TypeString RuntimeInitializeOnLoadMethodAttribute = new("RuntimeInitializeOnLoadMethodAttribute", Namespace);
     }
 }