+++ Roslyn~/Coimbra.Services.Events.Roslyn/CoimbraServicesEventsTypes.cs
@@ -8,7 +8,13 @@ namespace Coimbra.Services.Events.Roslyn
 
         public static readonly TypeString AllowEventServiceUsageForAttribute = new("AllowEventServiceUsageForAttribute", Namespace);
 
+        public static readonly TypeString EventContextHandlerStruct = new("EventContextHandler", Namespace);
+
+        public static readonly TypeString EventHandleStruct = new("EventHandle", Namespace);
+
         public static readonly TypeString EventInterface = new("IEvent", Namespace);
+
+        public static readonly TypeString EventRelevancyChangedHandlerDelegate = new("EventRelevancyChangedHandler", Namespace);
 
         public static readonly TypeString EventServiceInterface = new("IEventService", Namespace);